+++ development-plugin/Admin/class-admin.php
@@ -42,9 +42,9 @@ class Admin {
 
 		$version = time();
 
-		$url = WP_PLUGIN_URL . '/bh-wp-logger-test-plugin/Admin/css/bh-wp-logger-test-plugin-admin.css';
+		$url = WP_PLUGIN_URL . '/bh-wp-logger-development-plugin/Admin/css/bh-wp-logger-development-plugin-admin.css';
 
-		wp_enqueue_style( 'bh-wp-logger-test-plugin', $url, array(), $version, 'all' );
+		wp_enqueue_style( 'bh-wp-logger-development-plugin', $url, array(), $version, 'all' );
 	}
 
 	/**
@@ -56,9 +56,9 @@ class Admin {
 
 		$version = time();
 
-		$url = WP_PLUGIN_URL . '/bh-wp-logger-test-plugin/Admin/js/bh-wp-logger-test-plugin-admin.js';
+		$url = WP_PLUGIN_URL . '/bh-wp-logger-development-plugin/Admin/js/bh-wp-logger-development-plugin-admin.js';
 
-		wp_enqueue_script( 'bh-wp-logger-test-plugin', $url, array( 'jquery' ), $version, true );
+		wp_enqueue_script( 'bh-wp-logger-development-plugin', $url, array( 'jquery' ), $version, true );
 	}
 
 	/**
@@ -105,6 +105,6 @@ class Admin {
 			$wp_debug_log = true === WP_DEBUG_LOG ? 'enabled' : WP_DEBUG_LOG;
 		}
 
-		include wp_normalize_path( __DIR__ . '/partials/bh-wp-logger-test-plugin-admin-display.php' );
+		include wp_normalize_path( __DIR__ . '/partials/bh-wp-logger-development-plugin-admin-display.php' );
 	}
 }