+++ tests/bootstrap.php
@@ -4,8 +4,8 @@
  */
 
 $GLOBALS['project_root_dir']   = $project_root_dir  = dirname( __DIR__, 1 );
-$GLOBALS['plugin_root_dir']    = $plugin_root_dir   = $project_root_dir . '/test-plugin';
-$GLOBALS['plugin_name']        = $plugin_name       = 'bh-wp-logger-test-plugin';
+$GLOBALS['plugin_root_dir']    = $plugin_root_dir   = $project_root_dir . '/development-plugin';
+$GLOBALS['plugin_name']        = $plugin_name       = 'bh-wp-logger-development-plugin';
 $GLOBALS['plugin_name_php']    = $plugin_name_php   = $plugin_name . '.php';
 $GLOBALS['plugin_path_php']    = $plugin_root_dir . '/' . $plugin_name_php;
 $GLOBALS['plugin_basename']    = $plugin_name . '/' . $plugin_name_php;